refactor: improve map type safety and extract to custom hooks
- Remove all `any` casts, add CityWithGeometryAndCounts type
- Extract magic numbers to MAP_HEATMAP_CONFIG constants
- Reduce main component from 714→249 lines (65% reduction)
- Create 6 custom hooks (useMapOptions, useMapPersonalization, etc.)
- Extract MapPopup component, clean up unused imports
